requirements.txt
Flask
cryptography
python-dotenv